Add dual LED support and MONITOR0 pin toggling to MCU
This MCU code update has not been tested yet!

This will toggle any excitations off during non-integration times of the PYTHON480. If DUAL_LED_MINISCOPE is selected at the top of main.c the MCU will switch between LED1 and LED2 on subsequent frames.

An update to the Miniscope DAQ firmware needs to also be made to enable adjustable control of the second LED intensity. This has not been done yet.